NoImplicityAny and Strict Null Checks

+++ src/github/pullRequestManager.ts
@@ -8,14 +8,14 @@ import * as Github from '@octokit/rest';
 import { CredentialStore } from './credentials';
 import { Comment } from '../common/comment';
 import { Remote, parseRepositoryRemotes } from '../common/remote';
-import { TimelineEvent, EventType, isReviewEvent, isCommitEvent } from '../common/timelineEvent';
+import { TimelineEvent, EventType, ReviewEvent as CommonReviewEvent, isReviewEvent, isCommitEvent } from '../common/timelineEvent';
 import { GitHubRepository, PULL_REQUEST_PAGE_SIZE } from './githubRepository';
 import { IPullRequestsPagingOptions, PRType, ReviewEvent, ITelemetry, IPullRequestEditData, PullRequest, IRawFileChange } from './interface';
 import { PullRequestGitHelper } from './pullRequestGitHelper';
 import { PullRequestModel } from './pullRequestModel';
 import { GitHubManager } from '../authentication/githubServer';
 import { formatError, uniqBy, Predicate } from '../common/utils';
-import { Repository, RefType, UpstreamRef, Branch } from '../typings/git';
+import { Repository, RefType, UpstreamRef } from '../typings/git';
 import Logger from '../common/logger';
 import { convertRESTPullRequestToRawPullRequest, convertPullRequestsGetCommentsResponseItemToComment, convertIssuesCreateCommentResponseToComment, parseGraphQLTimelineEvents, convertRESTTimelineEvents, parseGraphQLComment } from './utils';
 import { PendingReviewIdResponse, TimelineEventsResponse, PullRequestCommentsResponse, AddCommentResponse, SubmitReviewResponse, DeleteReviewResponse } from './graphql';
@@ -23,7 +23,7 @@ const queries = require('./queries.gql');
 
 interface PageInformation {
 	pullRequestPage: number;
-	hasMorePages: boolean;
+	hasMorePages: boolean | null;
 }
 
 interface RestErrorResult {
@@ -134,11 +134,11 @@ export class PullRequestManager {
 		}
 	}
 
-	get activePullRequest() {
+	get activePullRequest(): (PullRequestModel | undefined) {
 		return this._activePullRequest;
 	}
 
-	set activePullRequest(pullRequest: PullRequestModel) {
+	set activePullRequest(pullRequest: (PullRequestModel | undefined)) {
 		this._activePullRequest = pullRequest;
 		this._onDidChangeActivePullRequest.fire();
 	}
@@ -159,14 +159,14 @@ export class PullRequestManager {
 		Logger.debug('update repositories', PullRequestManager.ID);
 		const remotes = parseRepositoryRemotes(this.repository);
 		const potentialRemotes = remotes.filter(remote => remote.host);
-		let gitHubRemotes = await Promise.all(potentialRemotes.map(remote => this._githubManager.isGitHub(remote.gitProtocol.normalizeUri())))
+		let gitHubRemotes = await Promise.all(potentialRemotes.map(remote => this._githubManager.isGitHub(remote.gitProtocol.normalizeUri()!)))
 			.then(results => potentialRemotes.filter((_, index, __) => results[index]))
 			.catch(e => {
 				Logger.appendLine(`Resolving GitHub remotes failed: ${formatError(e)}`);
 				vscode.window.showErrorMessage(`Resolving GitHub remotes failed: ${formatError(e)}`);
 				return [];
 			});
-		gitHubRemotes = uniqBy(gitHubRemotes, remote => remote.gitProtocol.normalizeUri().toString());
+		gitHubRemotes = uniqBy(gitHubRemotes, remote => remote.gitProtocol.normalizeUri()!.toString());
 
 		if (gitHubRemotes.length) {
 			await vscode.commands.executeCommand('setContext', 'github:hasGitHubRemotes', true);
@@ -178,7 +178,7 @@ export class PullRequestManager {
 		}
 
 		let serverAuthPromises = [];
-		for (let server of uniqBy(gitHubRemotes, remote => remote.gitProtocol.normalizeUri().authority)) {
+		for (let server of uniqBy(gitHubRemotes, remote => remote.gitProtocol.normalizeUri()!.authority)) {
 			serverAuthPromises.push(this._credentialStore.hasOctokit(server).then(authd => {
 				if (!authd) {
 					this._credentialStore.loginWithConfirmation(server);
@@ -191,8 +191,8 @@ export class PullRequestManager {
 			Logger.appendLine(`serverAuthPromises failed: ${formatError(e)}`);
 		});
 
-		let repositories = [];
-		let resolveRemotePromises = [];
+		let repositories: GitHubRepository[] = [];
+		let resolveRemotePromises: Promise<void>[] = [];
 		let userCreatedRemoteNames = this._includeRemotes === IncludeRemote.All ? (gitHubRemotes as Remote[]) : await PullRequestGitHelper.getUserCreatedRemotes(this.repository, (gitHubRemotes as Remote[]));
 
 		userCreatedRemoteNames.forEach(remote => {
@@ -247,8 +247,8 @@ export class PullRequestManager {
 		}
 
 		const localBranches = this.repository.state.refs
-			.filter(r => r.type === RefType.Head && r.name)
-			.map(r => r.name);
+			.filter(r => r.type === RefType.Head && r.name !== undefined)
+			.map(r => r.name!);
 
 		const promises = localBranches.map(async localBranchName => {
 			const matchingPRMetadata = await PullRequestGitHelper.getMatchingPullRequestMetadataForBranch(this.repository, localBranchName);
@@ -258,7 +258,7 @@ export class PullRequestManager {
 				const githubRepo = githubRepositories.find(repo => repo.remote.owner.toLocaleLowerCase() === owner.toLocaleLowerCase());
 
 				if (githubRepo) {
-					const pullRequest: PullRequestModel = await githubRepo.getPullRequest(prNumber);
+					const pullRequest: PullRequestModel | undefined = await githubRepo.getPullRequest(prNumber);
 
 					if (pullRequest) {
 						pullRequest.localBranchName = localBranchName;
@@ -271,14 +271,17 @@ export class PullRequestManager {
 		});
 
 		return Promise.all(promises).then(values => {
-			return values.filter(value => value !== null);
+			return values.filter(value => value !== null) as PullRequestModel[];
 		});
 	}
 
 	async deleteLocalPullRequest(pullRequest: PullRequestModel, force?: boolean): Promise<void> {
+		if (!pullRequest.localBranchName) {
+			return;
+		}
 		await this.repository.deleteBranch(pullRequest.localBranchName, force);
 
-		let remoteName: string = null;
+		let remoteName: string | undefined = undefined;
 		try {
 			remoteName = await this.repository.getConfig(`branch.${pullRequest.localBranchName}.remote`);
 		} catch (e) {}
@@ -317,7 +320,10 @@ export class PullRequestManager {
 			}
 		}
 
-		githubRepositories = githubRepositories.filter(repo => this._repositoryPageInformation.get(repo.remote.url.toString()).hasMorePages !== false);
+		githubRepositories = githubRepositories.filter(repo => {
+			let info = this._repositoryPageInformation.get(repo.remote.url.toString());
+			return info && info.hasMorePages !== false;
+		});
 
 		let pullRequests: PullRequestModel[] = [];
 		let numPullRequests = 0;
@@ -334,7 +340,7 @@ export class PullRequestManager {
 			const shouldLoad = this._includeRemotes === IncludeRemote.All || !(await PullRequestGitHelper.isRemoteCreatedForPullRequest(this.repository, remote));
 			if (shouldLoad) {
 				const pageInformation = this._repositoryPageInformation.get(githubRepository.remote.url.toString());
-				while (numPullRequests < PULL_REQUEST_PAGE_SIZE && pageInformation.hasMorePages !== false) {
+				while (numPullRequests < PULL_REQUEST_PAGE_SIZE && pageInformation && pageInformation.hasMorePages !== false) {
 					const pullRequestData = await githubRepository.getPullRequests(type, pageInformation.pullRequestPage);
 					if (!pullRequestData) {
 						break;
@@ -353,7 +359,10 @@ export class PullRequestManager {
 	}
 
 	public mayHaveMorePages(): boolean {
-		return this._githubRepositories.some(repo => this._repositoryPageInformation.get(repo.remote.url.toString()).hasMorePages !== false);
+		return this._githubRepositories.some(repo => {
+			let info = this._repositoryPageInformation.get(repo.remote.url.toString());
+			return !!(info && info.hasMorePages !== false);
+		});
 	}
 
 	async getStatusChecks(pullRequest: PullRequestModel): Promise<Github.ReposGetCombinedStatusForRefResponse> {
@@ -388,11 +397,12 @@ export class PullRequestManager {
 			});
 
 			const comments = data.repository.pullRequest.reviews.nodes
-				.map(node => node.comments.nodes.map(comment => parseGraphQLComment(comment), remote))
-				.reduce((prev, curr) => curr = prev.concat(curr), []);
+				.map((node: any) => node.comments.nodes.map((comment: any) => parseGraphQLComment(comment), remote))
+				.reduce((prev: any, curr: any) => curr = prev.concat(curr), []);
 			return comments;
 		} catch (e) {
 			Logger.appendLine(`Failed to get pull request review comments: ${formatError(e)}`);
+			return [];
 		}
 	}
 
@@ -463,7 +473,7 @@ export class PullRequestManager {
 					number: pullRequest.prNumber
 				}
 			});
-			ret = data.repository.pullRequest.timeline.edges.map(edge => edge.node);
+			ret = data.repository.pullRequest.timeline.edges.map((edge: any) => edge.node);
 			let events = parseGraphQLTimelineEvents(ret);
 			await this.addReviewTimelineEventComments(pullRequest, events);
 			return events;
@@ -507,7 +517,7 @@ export class PullRequestManager {
 		return this.addCommentPermissions(convertIssuesCreateCommentResponseToComment(promise.data), remote);
 	}
 
-	async createCommentReply(pullRequest: PullRequestModel, body: string, reply_to: Comment): Promise<Comment> {
+	async createCommentReply(pullRequest: PullRequestModel, body: string, reply_to: Comment): Promise<Comment | undefined> {
 		const pendingReviewId = await this.getPendingReviewId(pullRequest);
 		if (pendingReviewId) {
 			return this.addCommentToPendingReview(pullRequest, pendingReviewId, body, { inReplyTo: reply_to.graphNodeId });
@@ -550,7 +560,7 @@ export class PullRequestManager {
 
 	async startReview(pullRequest: PullRequestModel): Promise<void> {
 		const { mutate } = await pullRequest.githubRepository.ensure();
-		return mutate<void>({
+		await mutate<void>({
 			mutation: queries.StartReview,
 			variables: {
 				input: {
@@ -561,15 +571,21 @@ export class PullRequestManager {
 		}).then(x => x.data).catch(e => {
 			Logger.appendLine(`Failed to start review: ${e.message}`);
 		});
+
+		return;
 	}
 
 	async inDraftMode(pullRequest: PullRequestModel): Promise<boolean> {
 		return !!await this.getPendingReviewId(pullRequest);
 	}
 
-	async getPendingReviewId(pullRequest = this._activePullRequest): Promise<string | null> {
+	async getPendingReviewId(pullRequest = this._activePullRequest): Promise<string | undefined> {
+		if (!pullRequest) {
+			return undefined;
+		}
+
 		if (!pullRequest.githubRepository.supportsGraphQl()) {
-			return null;
+			return;
 		}
 
 		const { query, octokit } = await pullRequest.githubRepository.ensure();
@@ -584,7 +600,7 @@ export class PullRequestManager {
 			});
 			return data.node.reviews.nodes[0].id;
 		} catch (error) {
-			return null;
+			return;
 		}
 	}
 
@@ -605,7 +621,7 @@ export class PullRequestManager {
 		return parseGraphQLComment(comment);
 	}
 
-	async createComment(pullRequest: PullRequestModel, body: string, path: string, position: number): Promise<Comment> {
+	async createComment(pullRequest: PullRequestModel, body: string, path: string, position: number): Promise<Comment | undefined> {
 		const pendingReviewId = await this.getPendingReviewId(pullRequest as PullRequestModel);
 		if (pendingReviewId) {
 			return this.addCommentToPendingReview(pullRequest as PullRequestModel, pendingReviewId, body, { path, position });
@@ -657,8 +673,12 @@ export class PullRequestManager {
 
 	async getHeadCommitMessage(): Promise<string> {
 		const { repository } = this;
-		const { message } = await repository.getCommit(repository.state.HEAD.commit);
-		return message;
+		if (repository.state.HEAD && repository.state.HEAD.commit) {
+			const { message } = await repository.getCommit(repository.state.HEAD.commit);
+			return message;
+		}
+
+		return '';
 	}
 
 	get origin(): GitHubRepository {
@@ -674,7 +694,7 @@ export class PullRequestManager {
 				// No GitHubRepository? We currently won't try pushing elsewhere,
 				// so fail.
 				throw new BadUpstreamError(
-					this.repository.state.HEAD.name,
+					this.repository.state.HEAD!.name!,
 					upstreamRef,
 					'is not a GitHub repo');
 			}
@@ -701,7 +721,7 @@ export class PullRequestManager {
 		return HEAD && HEAD.upstream;
 	}
 
-	async createPullRequest(params: Github.PullRequestsCreateParams): Promise<PullRequestModel> {
+	async createPullRequest(params: Github.PullRequestsCreateParams): Promise<PullRequestModel | undefined> {
 		try {
 			const repo = this._githubRepositories.find(r => r.remote.owner === params.owner && r.remote.repositoryName === params.repo);
 			if (!repo) {
@@ -732,7 +752,6 @@ export class PullRequestManager {
 		} catch (e) {
 			Logger.appendLine(`GitHubRepository> Creating pull requests failed: ${e}`);
 			vscode.window.showWarningMessage(`Creating pull requests for '${params.head}' failed: ${formatError(e)}`);
-			return null;
 		}
 	}
 
@@ -804,7 +823,7 @@ export class PullRequestManager {
 	}
 
 	private addCommentPermissions(rawComment: Comment, remote: Remote): Comment {
-		const isCurrentUser = this._credentialStore.isCurrentUser(rawComment.user.login, remote);
+		const isCurrentUser = this._credentialStore.isCurrentUser(rawComment.user!.login, remote);
 		const notOutdated = rawComment.position !== null;
 		rawComment.canEdit = isCurrentUser && notOutdated;
 		rawComment.canDelete = isCurrentUser && notOutdated;
@@ -986,13 +1005,13 @@ export class PullRequestManager {
 
 	//#region Git related APIs
 
-	async resolvePullRequest(owner: string, repositoryName: string, pullReuqestNumber: number): Promise<PullRequestModel> {
+	async resolvePullRequest(owner: string, repositoryName: string, pullReuqestNumber: number): Promise<PullRequestModel | undefined> {
 		const githubRepo = this._githubRepositories.find(repo =>
 			repo.remote.owner.toLowerCase() === owner.toLowerCase() && repo.remote.repositoryName.toLowerCase() === repositoryName.toLowerCase()
 		);
 
 		if (!githubRepo) {
-			return null;
+			return;
 		}
 
 		const pr = await githubRepo.getPullRequest(pullReuqestNumber);
@@ -1000,12 +1019,11 @@ export class PullRequestManager {
 	}
 
 	async getMatchingPullRequestMetadataForBranch() {
-		if (!this.repository || !this.repository.state.HEAD) {
+		if (!this.repository || !this.repository.state.HEAD || !this.repository.state.HEAD.name) {
 			return null;
 		}
 
-		const HEAD = this.repository.state.HEAD;
-		let matchingPullRequestMetadata = await PullRequestGitHelper.getMatchingPullRequestMetadataForBranch(this.repository, HEAD.name);
+		let matchingPullRequestMetadata = await PullRequestGitHelper.getMatchingPullRequestMetadataForBranch(this.repository, this.repository.state.HEAD.name);
 		return matchingPullRequestMetadata;
 	}
 
@@ -1015,22 +1033,6 @@ export class PullRequestManager {
 
 	async fetchAndCheckout(pullRequest: PullRequestModel): Promise<void> {
 		await PullRequestGitHelper.fetchAndCheckout(this.repository, this._githubRepositories, pullRequest);
-	}
-
-	async getBranch(remote: Remote, branchName: string): Promise<Branch> {
-		let githubRepository = this.findRepo(byRemoteName(remote.remoteName));
-		if (githubRepository) {
-			let githubBranch = await githubRepository.getBranch(branchName);
-
-			if (githubBranch) {
-				return {
-					name: githubBranch.name,
-					type: RefType.RemoteHead
-				};
-			}
-		}
-
-		return null;
 	}
 
 	async checkout(branchName: string): Promise<void> {
@@ -1071,12 +1073,12 @@ export class PullRequestManager {
 			index[evt.id] = evt;
 			evt.comments = [];
 			return index;
-		}, {});
+		}, {} as { [key: number]: CommonReviewEvent });
 
 		const commentsById = reviewComments.reduce((index, evt) => {
 			index[evt.id] = evt;
 			return index;
-		}, {});
+		}, {} as { [key: number]: CommentNode });
 
 		const roots: CommentNode[] = [];
 		let i = reviewComments.length; while (i --> 0) {
@@ -1091,7 +1093,7 @@ export class PullRequestManager {
 		}
 
 		roots.forEach(c => {
-			const review = reviewEventsById[c.pullRequestReviewId];
+			const review = reviewEventsById[c.pullRequestReviewId!];
 			review.comments = review.comments.concat(c).concat(c.childComments || []);
 		});
